% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bentobox.R
\docType{package}
\name{bentobox-package}
\alias{bentobox}
\alias{bentobox-package}
\title{bentobox: Built Envionment Network objects Modeling Toolbox}
\description{
bentobox provides a toolbox of data structures and modeling strategies for working with and analyzing built environment data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/apeterson91/bentobox/issues}
}

}
\author{
\strong{Maintainer}: Adam Peterson \email{atpvyc@umich.edu} (\href{https://orcid.org/-14945}{ORCID})

}
\keyword{internal}
