% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{bentobox_conflicts}
\alias{bentobox_conflicts}
\title{Conflicts between the bentobox and other packages}
\usage{
bentobox_conflicts()
}
\description{
This function lists all the conflicts between packages in the bentobox
and other packages that you have loaded.
}
\examples{
bentobox_conflicts()
}
