% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bentobox_packages}
\alias{bentobox_packages}
\title{List all packages in bentobox}
\usage{
bentobox_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include bentobox in the list?}
}
\description{
List all packages in bentobox
}
\examples{
bentobox_packages()
}
