% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorby.R
\name{colorby}
\alias{colorby}
\title{Handle BentoBox color scaling parameters}
\usage{
colorby(column, range = NULL)
}
\arguments{
\item{column}{String specifying name of data column to scale colors by.}

\item{range}{A numeric vector specifying the range of values to
apply a color scale to.}
}
\value{
Returns a "\code{bb_colorby}" object.
}
\description{
\code{colorby} should be used to create a set of parameters
that specify color scaling for the functions \code{bb_plotPairs},
\code{bb_plotPairsArches}, and \code{bb_plotRanges}.
}
\examples{
## Load paired ranges data in BEDPE format
library(BentoBoxData)
data("bb_bedpeData")

## Add a length column
bb_bedpeData$length <- (bb_bedpeData$start2 - bb_bedpeData$start1) / 1000

## Plot pairs with colorby object set for `length` column
bedpePlot <- bb_plotPairs(
    data = bb_bedpeData,
    chrom = "chr21",
    chromstart = 27900000, chromend = 30700000,
    assembly = "hg19",
    fill = colorRampPalette(c("dodgerblue2", "firebrick2")),
    colorby = colorby("length"),
    lwd = 2, spaceHeight = .7,
)
}
